Reorganize token pooling code (#2792)
Ref #2698
